requirements.txt
numpy
matplotlib
flask
google-auth-oauthlib
google-auth
requests
cryptography 